moved getClients()
