+++ K6/api/building_blocks/auth/authorizedParties/index.js
@@ -1,1 +1,1 @@
-export { GetAuthorizedParties, getClients } from "./getAuthorizedParties.js"
+export { GetAuthorizedParties } from "./getAuthorizedParties.js"
