+++ K6/api/tests/auth/authorizedParties/getAuthorizedPartiesForUser.js
@@ -1,5 +1,6 @@
 import { SharedArray } from 'k6/data';
-import { GetAuthorizedParties, getClients } from '../../../building_blocks/auth/authorizedParties/index.js';
+import { GetAuthorizedParties } from '../../../building_blocks/auth/authorizedParties/index.js';
+import { getClients } from './getClients.js';
 import { getItemFromList, readCsv } from '../../../../helpers.js';
 
 const includeAltinn2 = (__ENV.INCLUDE_ALTINN2 ?? 'true') === 'true';
